Override the menu script to allow for an empty secondary menu region
The header.js provided by localgov_base expects there to be child nodes of the secondary menu region and when there isn't the script errors. The script has been copied into the lgd_org_theme child theme and altered to allow for an empty region. This requires a couple of template overrides in order to override the JS as the base theme is using attach_library.